requirements_complete.txt
# COMPLETE Autohire Requirements - Multi-Agent with Advanced Orchestration
# Python 3.9+

# ============================================================================
# I. CORE AUTOMATION (Existing)
# ============================================================================
playwright==1.41.0
pyautogui==0.9.54
pillow==10.2.0
pywin32==306
python-docx==1.1.0
cryptography==42.0.1
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# ============================================================================
# II. ADVANCED AGENT ORCHESTRATION & CONTROL FLOW
# ============================================================================

# LangGraph - State-based Cyclic Agent Workflows
langgraph==0.2.45
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-openai==0.2.5
langchain-anthropic==0.2.5
langchain-google-genai==2.0.5

# Pydantic-AI - Structured Input/Output with Reflection
pydantic-ai==0.0.14
pydantic==2.9.2
pydantic-settings==2.5.2

# AutoGen - Multi-Agent Collaboration (Planner, Coder, Executor)
pyautogen==0.2.35

# ============================================================================
# III. ROBUST DESKTOP AUTOMATION (Better than PyAutoGUI alone)
# ============================================================================

# Pywinauto - Windows GUI Automation (Element-based, not coordinate-based)
pywinauto==0.6.8
comtypes==1.4.11

# AutoHotkey Python Wrapper
ahk==1.7.4  # Python wrapper for AutoHotkey

# Screen Capture & Computer Vision
mss==9.0.1                    # Fast screenshot capture
pytesseract==0.3.13           # OCR for text recognition
opencv-python==4.10.0.84      # Computer vision
easyocr==1.7.2               # Alternative OCR (more accurate)

# VLM-based Screen Analysis (Vision-Language Models)
# Uses GPT-4-Vision or Claude-Vision via OpenAI/Anthropic APIs (no extra package)

# ============================================================================
# IV. OBSERVABILITY, DEBUGGING, AND EVALUATION
# ============================================================================

# Langfuse - Open-Source LLM Observability & Tracing
langfuse==2.52.3
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0

# Phoenix (Arize) - LLM and Embedding Observability
arize-phoenix==5.8.3
openinference-instrumentation-langchain==0.1.29

# LangSmith (Commercial, but has free tier)
langsmith==0.1.147

# ============================================================================
# V. ADVANCED KNOWLEDGE RETRIEVAL (RAG & Knowledge Graphs)
# ============================================================================

# LlamaIndex - RAG-focused Agents and Data Frameworks
llama-index==0.11.20
llama-index-core==0.11.20
llama-index-llms-openai==0.2.14
llama-index-llms-anthropic==0.2.5
llama-index-embeddings-openai==0.2.5
llama-index-vector-stores-chroma==0.2.0
llama-index-graph-stores-neo4j==0.3.0

# ChromaDB - Vector Store for Long-Term Memory
chromadb==1.0.21
sentence-transformers==3.3.1

# Neo4j - Knowledge Graph Database (Optional but powerful)
neo4j==5.25.0
py2neo==2021.2.4

# Advanced RAG Techniques
ragas==0.2.7                  # RAG evaluation framework
guardrails-ai==0.5.14         # Structured output validation

# ============================================================================
# VI. LLM PROVIDERS
# ============================================================================

# OpenAI (GPT-4o, GPT-4-Vision)
openai==1.68.2

# Anthropic (Claude 3.5 Sonnet, Claude-Vision)
anthropic==0.39.0

# Google (Gemini Pro - FREE tier available)
google-generativeai==0.8.3

# ============================================================================
# VII. SUPPORTING LIBRARIES
# ============================================================================

# Utilities
httpx==0.27.2
tenacity==9.0.0               # Retry logic
tiktoken==0.8.0               # Token counting
jinja2==3.1.4                 # Templating
PyYAML==6.0.2                 # Config files

# Database
sqlalchemy==2.0.36
alembic==1.13.3               # Database migrations
psycopg2-binary==2.9.10       # PostgreSQL driver

# Logging & Monitoring
loguru==0.7.2                 # Better logging
rich==13.8.1                  # Beautiful terminal output
tqdm==4.67.1                  # Progress bars

# Data Processing
numpy==1.26.4

# Testing & Quality
pytest==8.3.3
pytest-asyncio==0.24.0

# Background Tasks
celery==5.3.4
redis==5.2.0

# ============================================================================
# VIII. EXPERIMENTAL / RESEARCH FRAMEWORKS
# ============================================================================

# Self-Operating Computer Framework (install from GitHub if needed)
# pip install git+https://github.com/OthersideAI/self-operating-computer

# Browser-Use (Advanced browser automation)
browser-use==0.5.5

# Crawl4AI (Advanced web scraping)
crawl4ai==0.7.2
